+++ app/cli.py
@@ -1,11 +1,26 @@
 import os
-from app import app
+from flask import Blueprint
 import click
 
-@app.cli.group()
+bp = Blueprint('cli', __name__, cli_group=None)
+
+
+@bp.cli.group()
 def translate():
     """Translation and localization commands."""
     pass
+
+
+@translate.command()
+@click.argument('lang')
+def init(lang):
+    """Initialize a new language."""
+    if os.system('pybabel extract -F babel.cfg -k _l -o messages.pot .'):
+        raise RuntimeError('extract command failed')
+    if os.system(
+            'pybabel init -i messages.pot -d app/translations -l ' + lang):
+        raise RuntimeError('init command failed')
+    os.remove('messages.pot')
 
 
 @translate.command()
@@ -23,15 +38,3 @@ def compile():
     """Compile all languages."""
     if os.system('pybabel compile -d app/translations'):
         raise RuntimeError('compile command failed')
-    
-
-@translate.command()
-@click.argument('lang')
-def init(lang):
-    """Initialize a new language."""
-    if os.system('pybabel extract -F babel.cfg -k _l -o messages.pot .'):
-        raise RuntimeError('extract command failed')
-    if os.system(
-            'pybabel init -i messages.pot -d app/translations -l ' + lang):
-        raise RuntimeError('init command failed')
-    os.remove('messages.pot')
