+++ app/translate.py
@@ -1,14 +1,14 @@
 import requests
+from flask import current_app
 from flask_babel import _
-from app import app
 
 def translate(text, source_language, dest_language):
-    if 'MS_TRANSLATOR_KEY' not in app.config or not app.config['MS_TRANSLATOR_KEY']:
+    if 'MS_TRANSLATOR_KEY' not in current_app.config or not current_app.config['MS_TRANSLATOR_KEY']:
         return _('Error: the translation service is not configured.')
 
     headers = {
-        'Ocp-Apim-Subscription-Key': app.config['MS_TRANSLATOR_KEY'],
-        'Ocp-Apim-Subscription-Region': app.config['MS_TRANSLATOR_REGION'],  # add this in config
+        'Ocp-Apim-Subscription-Key': current_app.config['MS_TRANSLATOR_KEY'],
+        'Ocp-Apim-Subscription-Region': current_app.config['MS_TRANSLATOR_REGION'],  # add this in config
         'Content-Type': 'application/json'
     }
