+++ microblog.py
@@ -1,7 +1,9 @@
 import sqlalchemy as sa
 import sqlalchemy.orm as so
-from app import app, db, cli
+from app import create_app, db
 from app.models import User, Post
+
+app = create_app()
 
 
 @app.shell_context_processor
